Implementing the interface in the other classes
We can export the interface and implement it in the User and Company classes. The purpose of this is to pinpoint where an error occurs if one of the other classes isn't meeting the interface requirements. The error would show up in the index.ts file but it is more helpful to know that the source of the error is the User class.